refactor(protocol): collapse epic.createChat 1.2 into the unreleased 1.1
`epic.createChat@1.2` exists on a false premise. Its own comment justifies the
new minor with:

    A NEW MINOR for the same reason v1.1 was one - `epic.createChat@1.1` is
    already in released hosts.

It is not. The newest released baseline (`host-v1.1.11` / `cli-v1.1.10`, all at
commit c785d864) reports `epic.createChat` at `latestMinor: 0` - verified
against the tag's own `registry.ts`, not merely its published surface asset.
Only `1.0` ever shipped, so `1.1` was still free to grow and `1.2` bought
nothing but a second upgrade hop.

Fold the cross-host fork owner hint into `1.1`:

- `createChatForkSourceAssistantBoundarySchema` gains `sourceOwnerUserId`
  (nullable, `.default(null)`), so the tagged precise-boundary arm carries the
  hint directly instead of a parallel `...V12` twin
- delete `createChatForkSourceAssistantBoundarySchemaV12`,
  `createChatForkSourceSchemaV12`, `createChatRequestSchemaV12` and their types
- delete `epicCreateChatV12` and `epicCreateChatUpgradeV11ToV12`; the surviving
  `epicCreateChatUpgradeV10ToV11` now fills the honest `null` itself, which is
  exactly what the deleted V11->V12 hop did
- registry line drops to `latestMinor: 1`

`epic.createChat@1.0` IS released, so the split between it and the tagged union
is real and stays: `createChatForkSourceSchema` is untouched, and the emitted
`1.0` wire schema is byte-identical to what this tree produced before the
change.

Verified: only `epic.createChat` moves in the re-dumped surface (beside the
`host.update.install` collapse in the preceding commit), the released-peer
compat gate is green against `host-v1.1.11`, and protocol compiles.

The dedicated `epic-create-chat-v12-fork-owner-hint` suite tested the 1.1<->1.2
split itself, so it is rewritten as `...-v11-...` against the collapsed line -
keeping every assertion that still has meaning (hint defaulting, empty-string
rejection, the `latest` arm's explicit requirement, the V10->V11 upgrade, the
registry shape) and dropping the ones about a version that no longer exists.
