+++ protocol/src/host/registry.ts
@@ -286,10 +286,8 @@ import {
   epicBatchUpdateRolesV10,
   epicCreateArtifactV10,
   epicCreateChatUpgradeV10ToV11,
-  epicCreateChatUpgradeV11ToV12,
   epicCreateChatV10,
   epicCreateChatV11,
-  epicCreateChatV12,
   epicCreateCommentThreadV10,
   epicCreateTuiAgentV10,
   epicCreateTuiAgentV11,
@@ -5561,25 +5559,21 @@ const HOST_RPC_REGISTRY_BASE_DEFINITION = {
   },
   "epic.createChat": {
     1: {
-      latestMinor: 2,
+      latestMinor: 1,
       versions: {
         0: {
           contract: epicCreateChatV10,
           upgradeFromPreviousVersion: null,
         },
         // v1.1: `forkSource` widened to name a latest-checkpoint boundary
-        // alongside the existing precise one (chat-sync-v2 ticket 34B1).
+        // alongside the existing precise one (chat-sync-v2 ticket 34B1), and
+        // the PRECISE-boundary variant carries the `sourceOwnerUserId` hint so
+        // a cross-host fork's target host can satisfy the cloud tier's owner
+        // check for a chat it holds no registry facts about (cross-host fork
+        // ticket A1).
         1: {
           contract: epicCreateChatV11,
           upgradeFromPreviousVersion: epicCreateChatUpgradeV10ToV11,
-        },
-        // v1.2: the PRECISE-boundary variant gains the `sourceOwnerUserId`
-        // hint v1.1 gave the latest-checkpoint one, so a cross-host fork's
-        // target host can satisfy the cloud tier's owner check for a chat it
-        // holds no registry facts about (cross-host fork ticket A1).
-        2: {
-          contract: epicCreateChatV12,
-          upgradeFromPreviousVersion: epicCreateChatUpgradeV11ToV12,
         },
       },
       downgradePathsFromLatest: {},
